requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Azure / OneLake
azure-identity==1.19.0
azure-storage-file-datalake==12.17.0

# AI / LLM
langchain==0.3.14
langchain-openai==0.2.14
langgraph==0.2.59
openai==1.58.1

# Vector Database
qdrant-client==1.12.1

# Data Processing
pandas==2.2.3
numpy==2.1.3
pyarrow==18.1.0
deltalake==0.22.3

# Caching
redis==5.2.1

# Utilities
structlog==24.4.0
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

